When a prepared statement uses different parameter types than what the statement was originally prepared for, the driver must replan the query for the new types.  When doing this in batches the driver was not correctly freeing old plans.
To cleanup prepared statements the driver records a reference to a
query object in response to the ParseComplete message.  The problem
in this case was that since the driver sends multiple Parse messages
before a Sync, when the driver receives the first ParseComplete
message the query object now contains a reference to the latest
parsed statement, not the first. So it was only cleaning up the
last parsed statement not all of them.

As reported by Peter Eisentraut.