The previous commit message is completely bogus.  I use the -F argument to specify a filename and I picked com.msg instead of oom.msg resulting in a random old message.  This is the correct log entry.
An OutOfMemoryError during fe<->be communication leaves the protocol
in an unknown state.  This prevents the caller from even closing the
connection.  Checking every allocation is a lot of work, but the
most likely error location is going to be when receiving large
result sets.  Put in checks around receiving row data to detect
memory exhaustion, keep the protocol in a known state, and report
the failure back to the user.

As reported by David Blasby.