Multiple calls to XAConnection.getConnection within the same user transaction ended up restarting the transaction on the server side as a result of manipulating the autocommit state.  When retrieving a Connection, we must pay attention to whether a user transaction is in progress when setting the autocommit state.
This only fixes the case where a close call is interspersed between
the getConnection calls.  Without the close call, the second
getConnection performs a silent rollback.  Even with the close, this
is dodgy behavior from the client and we expect it not to happen,
but it would be better if we actually forbade it.

Heikki Linnakangas